code/my-react-app/src/ClassMate-Backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.4
python-socketio==5.9.0
python-engineio==4.7.1
psycopg2-binary==2.9.9
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
